move RedirectCout to base/utilities.h
